+++ lib/an_lifecycle_cancellable.dart
@@ -1,6 +1,8 @@
 library lifecycle_cancellable;
 
 export 'dialog/context_ext.dart';
-export 'dialog/navigator_ext.dart';
 export 'dialog/dialog_ext.dart';
+export 'dialog/navigator_ext.dart';
+export 'lifecycle/lifecycle_async.dart';
 export 'lifecycle/lifecycle_ext.dart';
+export 'listenable/change_notifier_ext.dart';